+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryScreenUi.kt
@@ -16,4 +16,6 @@ interface PatientSummaryScreenUi {
   fun showNextAppointmentCard()
   fun hideNextAppointmentCard()
   fun showClinicalDecisionSupportAlert()
+  fun hideClinicalDecisionSupportAlert()
+  fun hideClinicalDecisionSupportAlertWithoutAnimation()
 }