+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryViewRenderer.kt
@@ -28,17 +28,23 @@ class PatientSummaryViewRenderer(
       }
 
       renderNextAppointmentCard(model)
-
-      val canShowClinicalDecisionSupportAlert = model.hasPatientRegistrationData == true && model.readyToRender()
-      if (canShowClinicalDecisionSupportAlert) {
-        clinicalDecisionSupportCallback.pass(model.isNewestBpEntryHigh == true, ::renderClinicalDecisionSupportAlert)
-      }
+      renderClinicalDecisionSupportAlert(model)
     }
   }
 
-  private fun renderClinicalDecisionSupportAlert(isNewestBpEntryHigh: Boolean) {
-    if (isNewestBpEntryHigh) {
-      ui.showClinicalDecisionSupportAlert()
+  private fun renderClinicalDecisionSupportAlert(model: PatientSummaryModel) {
+    val canShowClinicalDecisionSupportAlert = model.hasPatientRegistrationData == true && model.readyToRender()
+    if (!canShowClinicalDecisionSupportAlert) {
+      ui.hideClinicalDecisionSupportAlertWithoutAnimation()
+      return
+    }
+
+    clinicalDecisionSupportCallback.pass(model.isNewestBpEntryHigh == true) { isNewestBpEntryHigh ->
+      if (isNewestBpEntryHigh) {
+        ui.showClinicalDecisionSupportAlert()
+      } else {
+        ui.hideClinicalDecisionSupportAlert()
+      }
     }
   }