+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Bump coroutines to v1.6.0
 - Implemented the support for DrugStockReminder API
 - Show the clinical decision support banner in summary screen when newest BP entry is high for the patient
+- Hide clinical decision support alert banner when edit medicines are clicked
 
 ### Features